Move password class and fix regex
